# Change code generation strategy for ILEmitResolverBuilder