--- a/src/libraries/Microsoft.Extensions.DependencyInjection/src/ServiceLookup/CompiledServiceProviderEngine.cs
+++ b/src/libraries/Microsoft.Extensions.DependencyInjection/src/ServiceLookup/CompiledServiceProviderEngine.cs
@@ -18,6 +18,6 @@ namespace Microsoft.Extensions.DependencyInjection.ServiceLookup
             ResolverBuilder = new(provider);
         }
 
-        public override Func<ServiceProviderEngineScope, object> RealizeService(ServiceCallSite callSite) => ResolverBuilder.Build(callSite);
+        public override ServiceFactory RealizeService(ServiceCallSite callSite) => ResolverBuilder.Build(callSite);
     }
 }