--- a/src/libraries/Microsoft.Extensions.DependencyInjection/src/ServiceLookup/ILEmit/ILEmitResolverBuilderContext.cs
+++ b/src/libraries/Microsoft.Extensions.DependencyInjection/src/ServiceLookup/ILEmit/ILEmitResolverBuilderContext.cs
@@ -3,14 +3,17 @@
 
 using System;
 using System.Collections.Generic;
+using System.Reflection;
 using System.Reflection.Emit;
 
 namespace Microsoft.Extensions.DependencyInjection.ServiceLookup
 {
     internal sealed class ILEmitResolverBuilderContext
     {
+        public FieldInfo RuntimeContextField { get; set; }
         public ILGenerator Generator { get; set; }
         public List<object> Constants { get; set; }
         public List<Func<IServiceProvider, object>> Factories { get; set; }
+        public HashSet<Assembly> Assemblies { get; set; }
     }
 }